face_detection/requirements.txt
numpy
opencv-python
PyYAML
tqdm
Cython
ipython
tensorflow-gpu==2.1.0
dlib
h5py==2.10.0 